LayoutTests/fast/forms/checkValidity-001.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../js/resources/js-test-pre.js"></script>
</head>
<body>
<p id="description"></p>
<form method="get">
<fieldset name="victim"></fieldset>
<input name="victim" type="text" value="lorem ipsum"/>
<button name="victim">lorem ipsum</button>
<select name="victim"></select>
<textarea name="victim"></textarea>
<output name="victim"></output>
<object name="victim"></object>
<keygen name="victim"></keygen>
</form>
<div id="console"></div>
<script>
description("This test checks if checkValidity() returns correctly a true (meaning no error) result.");

shouldBe('document.getElementsByTagName("fieldset")[0].checkValidity()', 'true');
shouldBe('document.getElementsByTagName("input")[0].checkValidity()', 'true');
shouldBe('document.getElementsByTagName("button")[0].checkValidity()', 'true');
shouldBe('document.getElementsByTagName("select")[0].checkValidity()', 'true');
shouldBe('document.getElementsByTagName("textarea")[0].checkValidity()', 'true');
shouldBe('document.getElementsByTagName("output")[0].checkValidity()', 'true');
shouldBe('document.getElementsByTagName("object")[0].checkValidity()', 'true');
shouldBe('document.getElementsByTagName("keygen")[0].checkValidity()', 'true');
</script>
<script src="../js/resources/js-test-post.js"></script>
</body>
</html>
